Add image module
