+++ lmnet/executor/predict.py
@@ -22,7 +22,7 @@ import click
 import numpy as np
 import tensorflow as tf
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.utils import config as config_util
 from lmnet.utils.executor import search_restore_filename
 from lmnet.utils.json import JsonOutput, ImageFromJson
