+++ lmnet/lmnet/datasets/bdd100k.py
@@ -7,7 +7,7 @@ import glob
 import numpy as np
 import json
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import ObjectDetectionBase
