+++ lmnet/lmnet/datasets/delta_mark.py
@@ -20,7 +20,7 @@ import os.path
 import numpy as np
 import pandas as pd
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import Base, ObjectDetectionBase, StoragePathCustomizable
 from lmnet import data_processor
 from lmnet.utils.random import shuffle, train_test_split
