+++ lmnet/lmnet/datasets/ilsvrc_2012.py
@@ -20,7 +20,7 @@ import os.path
 import numpy as np
 import pandas as pd
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import Base
 from lmnet import data_processor
