+++ lmnet/lmnet/datasets/image_folder.py
@@ -21,7 +21,7 @@ import os.path
 
 import numpy as np
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import Base, StoragePathCustomizable
 from lmnet import data_processor
 from lmnet.utils.random import train_test_split
