+++ lmnet/lmnet/datasets/mscoco.py
@@ -19,7 +19,7 @@ import os.path
 import numpy as np
 from pycocotools.coco import COCO
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import SegmentationBase
 from lmnet.datasets.base import ObjectDetectionBase
