+++ lmnet/lmnet/datasets/open_images_v4.py
@@ -25,7 +25,7 @@ import re
 import numpy as np
 
 from lmnet import data_processor
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import Base, StoragePathCustomizable
 from lmnet.datasets.base import ObjectDetectionBase
 from lmnet.utils.random import train_test_split
