+++ lmnet/lmnet/datasets/pascalvoc_2007_2012.py
@@ -17,7 +17,7 @@ import functools
 
 import numpy as np
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import ObjectDetectionBase
 from lmnet.datasets.pascalvoc_2007 import Pascalvoc2007
 from lmnet.datasets.pascalvoc_2012 import Pascalvoc2012
