+++ lmnet/lmnet/datasets/pascalvoc_base.py
@@ -21,7 +21,7 @@ import numpy as np
 import pandas as pd
 import xml.etree.ElementTree as ET
 
-from lmnet.pre_processor import load_image
+from lmnet.utils.image import load_image
 from lmnet.datasets.base import ObjectDetectionBase
