+++ lmnet/tests/lmnet_tests/test_data_augmentor.py
@@ -20,12 +20,10 @@ import PIL.Image
 import PIL.ImageDraw
 import pytest
 
+from lmnet.utils.image import load_image
 from lmnet.datasets.lm_things_on_a_table import LmThingsOnATable
 from lmnet.datasets.dataset_iterator import DatasetIterator
-from lmnet.pre_processor import (
-    ResizeWithGtBoxes,
-    load_image
-)
+from lmnet.pre_processor import ResizeWithGtBoxes
 from lmnet.data_processor import (
     Sequence
 )
