+++ output_template/python/lmnet/utils/image.py
@@ -1,0 +1,1 @@
+lmnet/lmnet/utils/image.py
